+++ examplepatches/franticg33k/morphe-patches/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.2.2](https://github.com/franticg33k/morphe-patches/compare/v1.2.1...v1.2.2) (2026-07-13)
+
+### 🚀 Updated App Support
+
+* add Atlas Photo support (Unlock Pro + Remove License Check) ([914cf24](https://github.com/franticg33k/morphe-patches/commit/914cf2447c7ea03c9f0303a63c31c9847ace5696))
+
+## [1.2.2-dev.1](https://github.com/franticg33k/morphe-patches/compare/v1.2.1...v1.2.2-dev.1) (2026-07-13)
+
+### 🚀 Updated App Support
+
+* add Atlas Photo support (Unlock Pro + Remove License Check) ([914cf24](https://github.com/franticg33k/morphe-patches/commit/914cf2447c7ea03c9f0303a63c31c9847ace5696))
+
 ## [1.2.1](https://github.com/franticg33k/morphe-patches/compare/v1.2.0...v1.2.1) (2026-07-11)
 
 ### 🚀 Updated App Support